requirements.txt
anyio==4.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.9
DBSQLite==1.0.0
dnspython==2.7.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
pyparsing==3.2.0
pyTelegramBotAPI==4.25.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.6.0
sniffio==1.3.1
telebot==0.0.5
# Editable Git install with no remote (telegram_drive==0.1.0)
-e /Users/rb/Downloads/pythonProjects/telegram_drive
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
